format string fixups
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox@27500 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ branches/ircd-ratbox-cleanup/include/ircd.h
@@ -47,11 +47,11 @@ struct SetOptions
 
 struct Counter
 {
-	int oper;		/* Opers */
-	int total;		/* total clients */
-	int invisi;		/* invisible clients */
-	int max_loc;		/* MAX local clients */
-	int max_tot;		/* MAX global clients */
+	unsigned long oper;		/* Opers */
+	unsigned long total;		/* total clients */
+	unsigned long invisi;		/* invisible clients */
+	unsigned long max_loc;		/* MAX local clients */
+	unsigned long max_tot;		/* MAX global clients */
 	unsigned long totalrestartcount;	/* Total client count ever */
 };
